Fix bug with symlink serialization
Previously, we used `Path::exists()` to validate that the target path
exists before writing any NAR data. However, this doesn't work when
attempting to serialize a symlink because `path.exists()` follows
symlinks by default, meaning it returns `false` if the symlink is
broken. In our case, we would like to serialize the symlink itself
regardless of whether the target path it's pointing to actually exists.

This is fixed by replacing our use of `Path::exists()` with
`std::fs::symlink_metadata().is_ok()`, which behaves identically to the
former except it does not follow symlinks.