Set read-only permissions for files
This commit ensures that regular files and executables are both set
read-only (mode 0444 for files and 0555 for executables), just as they
are configured by Nix. We don't yet expose a setting for toggling this
on and off, but most likely this will be exposed as a per-archive
setting and not a per-entry setting (it doesn't make any sense
otherwise; if a parent directory is set to read-only, how will you make
its children writable?).