Ensure u64 to usize cast is lossless
This commit casts the `usize` values to `u64`, performs the modulus
operation, and then converts the small remainder back to `usize`. This
calculation is more likely to be lossless compared to the previous
method, which casts a potentially large value directly to `usize` before
performing the modulus which might truncate the value on 32-bit systems.